+++ test/lib/yenv.strict.spec.js
@@ -5,14 +5,14 @@ const sinon = require('sinon')
 const { UndefinedKeyError } = require('keyblade')
 const fixture = require('../_helpers/fixture')('importing')
 
-describe('strict mode', function() {
-  it('throws when keys are missing', function() {
+describe('strict mode', function () {
+  it('throws when keys are missing', function () {
     const logSpy = sinon.spy()
     const actual = yenv(fixture('importing.yaml'), {
       env: 'development',
       envObject: {},
       strict: true,
-      logBeforeThrow: (arg1, arg2) => logSpy(arg1, arg2)
+      logBeforeThrow: (arg1, arg2) => logSpy(arg1, arg2),
     })
     const err = catchError(() => actual.NONEXISTING)
     logSpy.should.have.been.called
@@ -27,7 +27,7 @@ describe('strict mode', function() {
       envObject: {},
       strict: true,
       logBeforeThrow: (arg1, arg2) => logSpy(arg1, arg2),
-      optionalKeys: ['OPTIONAL']
+      optionalKeys: ['OPTIONAL'],
     })
     expect(catchError(() => actual.NONEXISTING)).to.exist
     logSpy.should.have.been.called
@@ -42,7 +42,7 @@ describe('loose mode', () => {
       env: 'development',
       envObject: {},
       strict: false,
-      logBeforeThrow: (arg1, arg2) => logSpy(arg1, arg2)
+      logBeforeThrow: (arg1, arg2) => logSpy(arg1, arg2),
     })
     expect(actual.NONEXISTING).to.equal(undefined)
     logSpy.should.not.have.been.called