+++ CHANGELOG.md
@@ -2,9 +2,13 @@
 
 ## 2.0.0
 
-* **[BREAKING]**:Coerce to known type ([#12](https://github.com/jeffijoe/yenv/issues/12))
+* **[BREAKING]**:Coerce to known type ([#12](https://github.com/jeffijoe/yenv/issues/12)) (can be disabled)
 * **[BREAKING]**: `~import` no longer throws on missing files; introduced `~require` which does.
+* **[BREAKING]**: Typings now use a generic to describe the return object
+* **[BREAKING]**: Introduced `strict` mode which is on by default
 * Add CLI tool to print the environment variables.
+
+---
 
 ## 1.0.0
 
@@ -16,7 +20,7 @@
 
 ## 0.2.0
 
-* **FEATURE:** Added composition
+* **[FEATURE]:** Added composition
 
 ---